fix: fixed use of Length Validator for PhoneInput
