+++ src/hooks/validators/LengthValidator.ts
@@ -4,7 +4,7 @@
  * File Created: Friday, 14th August 2020 3:09:48 pm
  * Author: Esperanza Horn (esperanza@inventures.cl)
  * -----
- * Last Modified: Monday, 24th August 2020 3:34:04 pm
+ * Last Modified: Monday, 24th August 2020 3:53:38 pm
  * Modified By: Esperanza Horn (esperanza@inventures.cl)
  * -----
  * Copyright 2020 - 2020 Incrementa Ventures SpA. ALL RIGHTS RESERVED
@@ -16,11 +16,16 @@
 import { Validator } from './Validator';
 
 export class LengthValidator extends Validator {
-  constructor(errorMsg: string) {
+  maxLength: number;
+  constructor(errorMsg: string, maxLength: number) {
     super(errorMsg);
+    this.maxLength = maxLength;
   }
 
-  validate(input: string, maxLength: number) {
-    return Boolean(input.length === maxLength);
+  validate(input: string) {
+    return Boolean(input.length === this.maxLength);
+  }
+  get _tag() {
+    return this.errorMsg + `-${this.maxLength}`;
   }
 }
