+++ src/stories/1-Input.stories.tsx
@@ -4,8 +4,8 @@
  * File Created: Wednesday, 8th July 2020 1:55:18 am
  * Author: Gabriel Ulloa (gabriel@inventures.cl)
  * -----
- * Last Modified: Monday, 24th August 2020 2:56:42 pm
- * Modified By: Gabriel Ulloa (gabriel@inventures.cl)
+ * Last Modified: Monday, 24th August 2020 4:22:57 pm
+ * Modified By: Esperanza Horn (esperanza@inventures.cl)
  * -----
  * Copyright 2019 - 2020 Incrementa Ventures SpA. ALL RIGHTS RESERVED
  * Terms and conditions defined in license.txt
@@ -40,9 +40,6 @@ export type CountryType = {
   countryName: string;
   countryDigitLength: number;
   countryPrefix: number;
-};
-type InputforPhoneProps = {
-  possibleCountries: CountryType[];
 };
 
 export const Base = () => <Input />;
@@ -160,7 +157,7 @@ export const InputForPhone = () => {
       countryPrefix: 1,
     },
   ];
-    
+
   const classes = useStyles();
 
   //const { possibleCountries } = props;
@@ -188,20 +185,24 @@ export const InputForPhone = () => {
     validators: [
       required && new RequiredValidator(required),
       nonNumeric && new NumericValidator(nonNumeric),
-      incompleteNumber && new LengthValidator(incompleteNumber),
+      incompleteNumber &&
+        new LengthValidator(incompleteNumber, country.countryDigitLength),
     ].filter(Boolean) as Validator<string>[],
     debounceTime,
-    maxLength: country.countryDigitLength,
+    //maxLength: country.countryDigitLength,
   });
 
   // when change of country, update the country and run callback functionon useInput to update max length
   const handleChange = useCallback(
     (event: ChangeEvent<{ value: unknown }>) => {
-      const newCountry = event.target.value as CountryType;
+      const newCountry = possibleCountries.find(
+        (x) => x.countryName === event.target.value,
+      ) as CountryType;
+      //const newCountry = event.target.value as CountryType;
       setCountry(newCountry);
       updateMaxLength(newCountry.countryDigitLength);
     },
-    [updateMaxLength, setCountry],
+    [updateMaxLength, setCountry, possibleCountries],
   );
 
   const handleWrite = useCallback(
@@ -213,10 +214,10 @@ export const InputForPhone = () => {
 
   return (
     <FormControl className={classes.formControl}>
-      <Select value={country} onChange={handleChange}>
+      <Select value={country.countryName} onChange={handleChange}>
         {possibleCountries.map((item, index) => (
           // TODO: understand why it complains about using an object - yet compiles and works fine
-          <MenuItem key={index} value={item}>
+          <MenuItem key={index} value={item.countryName}>
             {item.countryName} (+{item.countryPrefix})
           </MenuItem>
         ))}
